requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
openai>=1.0.0
lxml>=4.9.0
